requirements.txt
numpy
pandas
scikit-learn
tensorflow
ipython
scipy
tqdm
matplotlib
music21
keras_metrics
